started analysis script + misc improvements to eval file

+++ src/evaluate.py
@@ -1,5 +1,6 @@
 from playwright.sync_api import sync_playwright
 import pickle
+import os
 
 with sync_playwright() as p:
 
@@ -9,14 +10,13 @@ with sync_playwright() as p:
     # Init cache
     cache_path = "performance-timings.pkl"
     performance_timings = {}
-    with open(cache_path, "wb") as file:
+
+    # Load file into context if it already exists
+    if os.path.exists(cache_path):
+        with open(cache_path, "wb") as file:
             pickle.dump(performance_timings, file)
 
     for divs in range(1000, 10100, 100):
-
-        # Load inprogress file
-        with open(cache_path, "rb") as file:
-            performance_timings = pickle.load(file)
 
         # Skip clause - Skip if divs category has already been successful
         if f"https://ramity.github.io/how-many-divs/{divs}-35px" in performance_timings:
